proposed_workflow/requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.4